most custom ops utilize some aten to executorch scalartype conversion. this test file standardized that across those files

Differential Revision: [D76464550](https://our.internmc.facebook.com/intern/diff/D76464550/)

cc SS-JIA manuelcandales cbilgin

[ghstack-poisoned]

+++ backends/vulkan/runtime/vk_api/Descriptor.cpp
@@ -32,8 +32,8 @@ BufferBindInfo::BufferBindInfo(
 
 BufferBindInfo::BufferBindInfo(
     const VulkanBuffer& buffer_p,
-    const uint32_t offset_p,
-    const uint32_t range_p)
+    const size_t offset_p,
+    const size_t range_p)
     : handle(buffer_p.handle()),
       offset(buffer_p.mem_offset() + offset_p),
       range(range_p) {
